Added chat subwindows
Added chat subwindows, with most of the preliminary work done on them. The "spellchecker" library has been added to the distribution, and spell checking works on the input widget on chat subwindows. Also added in a default subwindow width/height setting to the config file.

The "build_dist.py" script had the "spellchecker" library added to it.