Added a user file
Added a user system. For now, it will only remember the last server you connected to (and the settings required for that) and your nick, alternate nick, username, and realname preferences.

I also removed the option to load a custom named config file. Since config files are stored in user's home directory, I've decided to not try and manage multiple users on a single user account.