Changes to the connect dialog
Logo/banner no longer appears when the connect dialog is called while running. Also, the title of the dialog changes depending on *how* the dialog is called (it's different for being called from the merk.py script, for being called from the menu, and for being called because a connection failed).