Broke backwards compatibility for config files; new features
Changed the "notify on lost connection" setting to "notify on lost or failed connection"; if the automatic "new server" prompt is turned off, and this setting is turned on, the client will notify the user of the failed/lost connection with a QMsgBox.

New version number also features a "0." in the start, as originally intended (at least so far).