Fixed a bug
When hiding the main window from the systray menu, if both the window and the subwindow were maximized, the subwindow wouldn't be displayed properly. Now, MERK detects if both of these are true, and if they are, save the states when hiding; when showing again, it restores the states so the subwindow displays properly.