Menubar update
Main window menubar is now a tool bar that can be docked at the top of the window (the default) or the bottom of the window. Uses code shamelessly stole from ERK. It looks a lot better, and now maximized windows now appear completely, rather than having the "minimize, maximize, restore" buttons moved to the menubar.