Major update
Added a navigation tool: the "windowbar", a tool bar that lists all the currently open chat windows, and allows you to click to open that window. The windowbar has a bunch of settings (where it is located, entry text justification, whether to show icons or not, whether to show server windows or not, and whether to show the windowbar at all) that have all been added to the config file/module and the settings dialog. Also, I consolidated the "menubar" and the "windowbar" settings into one section of the settings dialog, "Toolbars".