Bugfix
The style editor wasn't working "automatically" and instantly when dark mode was turned on. This was due to how Qt handles palettes, especially when using assigned (ie, not stock) ones. This is fixed by using stylesheets to assign background and foreground colors rather than the palette system. The styler has been changed to use the new system, too. so it should work equally in dark mode and light mode.