Preperation for new feature
Added the ability to choose whether or not to include channel and private chat windows in the window bar. If nothing is selected to show, the windowbar is hidden. Added the settings to the config module and config file. TODO: need to add it to the settings dialog and the windowbar context menu