Major update!
Application palette colors are now dynamic! MERK now loads all palette colors from a configuration file. Two palettes are included: the default palette (the one that looks "normal"), and the palette for dark mode (which gets loaded if "dark mode" is turned on). Currently, I'm not planning on a way for users to edit palettes in the application, but that may change in the future. Palettes use HTML hex colors, so they're easy to swap out as needed.