Removed the "default" palette from the app
Instead of loading in a "default" palette, I'm just going to go ahead the use the app defaults. I've removed the default palette completely: it is no longer autogenerated by the styles system, and no longer loaded in the starter script. The "dark" palette is still used for dark mode, however.