Chat logs can now be optionally saved every 30 minutes
Added an option to the settings menu and settings dialog, as well as the config file. It's toggleable, and the save interval is saved in the config file. There is no way to choose the interval from the GUI, but it can be changed in the config file. The option is turned on by default.