New feature: windowbar entry context menus!
Each entry in the windowbar now gets their own context menu that allows you to do various things with the window, including closing them. This can be turned on and off in the settings dialog or the windowbar context menu. The windowbar is turning, slowly, into the Windows task manager :-P